requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.10
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
joblib==1.3.2
kombu==5.3.5
newsapi-python==0.2.7
nltk==3.6.5
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.26.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
